requirements.txt
alembic==1.7.6
certifi==2021.10.8
charset-normalizer==2.0.11
cycler==0.11.0
fonttools==4.29.1
hurry==1.1
hurry.filesize==0.9
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
install==1.3.5
kiwisolver==1.3.2
Mako==1.1.6
MarkupSafe==2.1.0
matplotlib==3.5.1
numpy==1.22.2
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
psutil==5.9.0
pyparsing==3.0.7
pyTelegramBotAPI==4.3.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.8
zipp==3.7.0
